Rename test folder to test-ui and create deno-tests folder
