+++ .vscode/settings.json
@@ -1,5 +1,9 @@
 {
   "editor.codeActionsOnSave": {
     "source.fixAll": "explicit"
-  }
+  },
+  "deno.enable": true,
+  "deno.config": "./deno.json",
+  "typescript.tsdk": null,
+  "deno.lint": true,
 }
