+++ deno-tests/main.ts
@@ -1,0 +1,5 @@
+import { assertEquals } from "https://deno.land/std@0.224.0/assert/mod.ts";
+
+Deno.test("add() works", () => {
+	assertEquals(2 + 3, 5);
+});
