+++ test-ui/src/components/NeighborComponent.vue
