+++ tsconfig.json
@@ -1,8 +1,5 @@
 {
 	"compilerOptions": {
-		"moduleResolution": "node",
-		"target": "es6",
-		"noEmit": true,
-		"allowImportingTsExtensions": true
+
 	}
 }
